fastapi_backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.7
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dateutil>=2.8.2
openpyxl>=3.1.0
et-xmlfile>=1.1.0
xlrd>=2.0.0
pymysql>=1.1.0
pyodbc>=4.0.0  # 用于 Access report1.mdb 迁移